external entities

+++ src/main/java/com/googlecode/mapperdao/plugins/ManyToManyPlugins.scala
@@ -218,23 +218,24 @@ class ManyToManyDeletePlugin(driver: Driver, mapperDao: MapperDaoImpl) extends B
 
 	override def idColumnValueContribution[PC, T](tpe: Type[PC, T], deleteConfig: DeleteConfig, events: Events, o: T with PC with Persisted, entityMap: UpdateEntityMap): List[(SimpleColumn, Any)] = Nil
 
-	override def before[PC, T](entity: Entity[PC, T], deleteConfig: DeleteConfig, events: Events, o: T with PC with Persisted, keyValues: List[(SimpleColumn, Any)], entityMap: UpdateEntityMap) = if (deleteConfig.propagate) {
-		val tpe = entity.tpe
-		tpe.table.manyToManyColumnInfos.filterNot(deleteConfig.skip(_)).foreach { ci =>
-			// execute before-delete-relationship events
-			events.executeBeforeDeleteRelationshipEvents(tpe, ci, o)
+	override def before[PC, T](entity: Entity[PC, T], deleteConfig: DeleteConfig, events: Events, o: T with PC with Persisted, keyValues: List[(SimpleColumn, Any)], entityMap: UpdateEntityMap) =
+		if (deleteConfig.propagate) {
+			val tpe = entity.tpe
+			tpe.table.manyToManyColumnInfos.filterNot(deleteConfig.skip(_)).foreach { ci =>
+				// execute before-delete-relationship events
+				events.executeBeforeDeleteRelationshipEvents(tpe, ci, o)
 
-			driver.doDeleteAllManyToManyRef(tpe, ci.column, keyValues.map(_._2))
+				driver.doDeleteAllManyToManyRef(tpe, ci.column, keyValues.map(_._2))
 
-			ci.column.foreign.entity match {
-				case ee: ExternalEntity[Any] =>
-					val fo = ci.columnToValue(o)
-					ee.manyToManyOnDeleteMap(ci.asInstanceOf[ColumnInfoTraversableManyToMany[_, _, Any]])(DeleteExternalManyToMany(deleteConfig, o, fo))
-				case _ =>
+				ci.column.foreign.entity match {
+					case ee: ExternalEntity[Any] =>
+						val fo = ci.columnToValue(o)
+						ee.manyToManyOnDeleteMap(ci.asInstanceOf[ColumnInfoTraversableManyToMany[_, _, Any]])(DeleteExternalManyToMany(deleteConfig, o, fo))
+					case _ =>
+				}
+
+				// execute after-delete-relationship events
+				events.executeAfterDeleteRelationshipEvents(tpe, ci, o)
 			}
-
-			// execute after-delete-relationship events
-			events.executeAfterDeleteRelationshipEvents(tpe, ci, o)
 		}
-	}
 }
